earring-leaf: use custom assembly for cycle counting
With this change it is possible to get a much higher resolution, which
in turn makes it much easier to run the chip at a lower frequency. Right
now it's running at around 250kHz, which uses a lot less power than 1MHz
as before. The PWM speed is still almost the same.

Not sure how much power this actually saves (the LEDs use the majority
of the power) but at least this mostly avoids the power vs PWM frequency
tradeoff.

This change causes code size to go up by about 30 bytes.