watch: add display.Sleep support
Turning off the LCD results in a massive decrease of current, by around
6mA. So it's definitely worthwhile.

Apart from that, this commit updates github.com/aykevl/badge, which has
some improvements in power consumption. Together, this results in a
standby current consumption of around 0.19mA. Not as low as it can go
(apparently it can go to 0.06mA), but a lot better than before. And
apparently also outcompeting InfiniTime by a fair margin.