watch: make timeout selection easier to touch
This turns the screen timeout selection into a two-column grid and makes
the labels much bigger, making it much easier to select one.

Perhaps this needs to be generalized, for now it works well enough.

Because tinygl got some optimizations/refactors at the same time, this
actually results in a binary size reduction.