earring-ring: restructure the animation code
Only update a few LEDs on each LED refresh, to avoid unnecessary
stuttering in the animation each 16 or so refreshes.

This resulted in slightly more current consumption, either because the
code got more efficient and the LEDs were on for a relatively longer
period, or else simply because the new logic resulted in brighter
lights.